requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.6
langchain-anthropic==0.2.4
langchain-community==0.3.4
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-text-splitters==0.3.1
langgraph==0.2.43
langgraph-checkpoint==2.0.2
langgraph-checkpoint-sqlite==2.0.1
langgraph-sdk==0.1.35
langsmith==0.1.139
marshmallow==3.23.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.10
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
yarl==1.17.1
